heterogenity fix

+++ candidates/quality.py
@@ -15,7 +15,7 @@ def candidate_quality(candidates, seq_to_candidates):
        
         for interval in candidate.mapped_sequences:
             seq_name = interval.data[1]
-            duplicates = int(seq_name.split("-")[1])
+            duplicates = float(seq_name.split("-")[1])
             candidate_hits += duplicates 
             total_hits += len(seq_to_candidates[seq_name].candidates) * duplicates
